move compression classes out of encoding package
